pyproject.toml
[project]
name = "transcribe-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.13.3",
    "aiosignal>=1.4.0",
    "annotated-types>=0.7.0",
    "anyio>=4.10.0",
    "asyncio>=4.0.0",
    "attrs>=25.3.0",
    "bidict>=0.23.1",
    "certifi>=2025.8.3",
    "cffi>=1.17.1",
    "charset-normalizer>=3.4.3",
    "click>=8.2.1",
    "cryptography>=45.0.6",
    "docutils>=0.22",
    "fastapi>=0.125.0",
    "frozenlist>=1.7.0",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "httptools>=0.6.4",
    "httpx>=0.28.1",
    "idna>=3.10",
    "ifaddr>=0.2.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "jwt>=1.4.0",
    "markdown2>=2.5.4",
    "markupsafe>=3.0.2",
    "multidict>=6.6.4",
    "narwhals>=2.7.0",
    "nicegui[redis]>=3.7.1",
    "orjson>=3.11.5",
    "packaging>=25.0",
    "plotly>=6.3.1",
    "propcache>=0.3.2",
    "pscript>=0.7.7",
    "pycparser>=2.22",
    "pydantic>=2.11.7",
    "pydantic-core>=2.35.0",
    "pydantic-settings>=2.10.1",
    "pygments>=2.19.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.1.1",
    "python-engineio>=4.12.2",
    "python-multipart>=0.0.22",
    "python-socketio>=5.14.0",
    "pytz>=2025.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "simple-websocket>=1.1.0",
    "sniffio>=1.3.1",
    "starlette>=0.49.1",
    "typing-extensions>=4.15.0",
    "typing-inspection>=0.4.1",
    "urllib3>=2.6.3",
    "uvicorn>=0.35.0",
    "uvloop>=0.21.0",
    "vbuild>=0.8.2",
    "watchfiles>=1.1.0",
    "websockets>=15.0.1",
    "wsproto>=1.2.0",
    "yarl>=1.20.1",
]
